refactor: move overlay-agent to examples, revert ash branding to agent-sh
- Move overlay-agent from built-in extensions to examples/extensions/
- Remove createFloatingPanel from ExtensionContext public API
  (FloatingPanel stays in src/utils/ for direct import)
- Revert docs prose from "ash" back to "agent-sh" (keep ash for
  backend identifier and package name prefixes like ash-mcp-bridge)
