+++ docs/usage.md
@@ -1,8 +1,8 @@
 # Usage Guide
 
-## Running ash
+## Running agent-sh
 
-The simplest way to run ash — just provide an API key and model:
+The simplest way to run agent-sh — just provide an API key and model:
 
 ```bash
 # Using environment variables
@@ -32,7 +32,7 @@ DEBUG=1 agent-sh --api-key "$KEY" --model gpt-4o
 
 ## Provider Examples
 
-ash works with any OpenAI-compatible API. Here are common configurations:
+agent-sh works with any OpenAI-compatible API. Here are common configurations:
 
 ### OpenAI
 
@@ -91,7 +91,15 @@ agent-sh --api-key dummy \
 
 ## Overlay Agent (Ctrl+\)
 
-Press **Ctrl+\\** anywhere — even inside vim, htop, or ssh — to summon the agent in a floating overlay. Type a query, and the agent's response streams into the panel.
+The overlay agent is an optional extension that lets you summon the agent from anywhere — even inside vim, htop, or ssh — by pressing **Ctrl+\\**. Type a query, and the agent's response streams into a floating panel.
+
+```bash
+# Install the extension
+cp examples/extensions/overlay-agent.ts ~/.agent-sh/extensions/
+
+# Or load directly
+agent-sh -e ./examples/extensions/overlay-agent.ts
+```
 
 The agent can read the terminal screen and send keystrokes via the built-in `terminal_read` and `terminal_keys` tools, enabling it to operate inside interactive programs.
 
@@ -102,7 +110,7 @@ Requires `@xterm/headless` for the dimmed background compositing:
 npm install @xterm/headless@5.5.0 @xterm/addon-serialize@0.13.0
 ```
 
-## Using ash as Your Default Shell
+## Using agent-sh as Your Default Shell
 
 Add to the end of your `~/.zshrc` or `~/.bashrc`:
 
@@ -116,7 +124,7 @@ The `AGENT_SH` guard prevents infinite recursion. The checks ensure it only laun
 
 ## Configuration
 
-ash stores settings and query history in `~/.agent-sh/`. Configure via `~/.agent-sh/settings.json` — all fields are optional with sensible defaults.
+agent-sh stores settings and query history in `~/.agent-sh/`. Configure via `~/.agent-sh/settings.json` — all fields are optional with sensible defaults.
 
 ### Provider Profiles
 
@@ -200,7 +208,7 @@ When cycling across providers (e.g. from OpenAI to Ollama), the API key and base
 
 ## Startup Banner
 
-On launch, ash displays a structured startup banner showing:
+On launch, agent-sh displays a structured startup banner showing:
 
 - **Backend** — which agent backend is active (`ash`, `claude-code`, `pi`, etc.)
 - **Model** — current model with provider in brackets (e.g. `gpt-4o [openai]`)
