+++ src/types.ts
@@ -4,7 +4,6 @@ import type { ColorPalette } from "./utils/palette.js";
 import type { BlockTransformOptions, FencedBlockTransformOptions } from "./utils/stream-transform.js";
 import type { ToolDefinition } from "./agent/types.js";
 import type { TerminalBuffer } from "./utils/terminal-buffer.js";
-import type { FloatingPanel, FloatingPanelConfig } from "./utils/floating-panel.js";
 import type { Compositor } from "./utils/compositor.js";
 
 export type { ContentBlock } from "./event-bus.js";
@@ -123,12 +122,6 @@ export interface ExtensionContext {
    * Lazily created on first access. Returns null if @xterm/headless is not installed.
    */
   terminalBuffer: TerminalBuffer | null;
-  /**
-   * Create a floating panel overlay. The panel composites a bordered box
-   * over the terminal with input routing, dimmed background, and
-   * handler-based customization.
-   */
-  createFloatingPanel: (config: FloatingPanelConfig) => FloatingPanel;
 
   // ── Compositor ─────────────────────────────────────────────────
   /**
